Cura 12.11.2

+++ package.sh
@@ -1,24 +1,23 @@
 #!/bin/bash
 
-# This script is to package the Cura package for Windows
-# This script should run under Windows with Cygwin.
+# This script is to package the Cura package for Windows/Linux and Mac OS X
+# This script should run under Linux and Mac OS X, as well as Windows with Cygwin.
 
 #############################
 # CONFIGURATION
 #############################
 
 ##Select the build target
-#BUILD_TARGET=${1:-all}
-BUILD_TARGET=win32
+BUILD_TARGET=${1:-all}
+#BUILD_TARGET=win32
 #BUILD_TARGET=linux
 #BUILD_TARGET=darwin
 #BUILD_TARGET=debian
 
-
+##Do we need to create the final archive
+ARCHIVE_FOR_DISTRIBUTION=1
 ##Which version name are we appending to the final archive
-export BUILD_NAME=13.11
-
-#Cura-13.11-win32
+export BUILD_NAME=13.11.2
 TARGET_DIR=Cura-${BUILD_NAME}-${BUILD_TARGET}
 
 ##Which versions of external programs to use
@@ -30,8 +29,7 @@ if [ -z ${CURA_ENGINE_REPO} ] ; then
 fi
 
 #############################
-# Funcion que verifica las herramientas
-# en el sistema
+# Support functions
 #############################
 function checkTool
 {
@@ -63,19 +61,117 @@ function extract
 	7z x -y $* >> log.txt
 }
 
+#############################
+# Actual build script
+#############################
+if [ "$BUILD_TARGET" = "all" ]; then
+	$0 win32
+	$0 linux
+	$0 darwin
+	exit
+fi
 
 # Change working directory to the directory the script is in
 # http://stackoverflow.com/a/246128
 SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-echo "\n".$SCRIPT_DIR."\n"
 cd $SCRIPT_DIR
 
 checkTool git "git: http://git-scm.com/"
 checkTool curl "curl: http://curl.haxx.se/"
-checkTool 7z "7zip: http://www.7-zip.org/"
-checkTool mingw32-make "mingw: http://www.mingw.org/"
+if [ $BUILD_TARGET = "win32" ]; then
+	#Check if we have 7zip, needed to extract and packup a bunch of packages for windows.
+	checkTool 7z "7zip: http://www.7-zip.org/"
+	checkTool mingw32-make "mingw: http://www.mingw.org/"
+fi
+#For building under MacOS we need gnutar instead of tar
+if [ -z `which gnutar` ]; then
+	TAR=tar
+else
+	TAR=gnutar
+fi
 
 
+#############################
+# Darwin
+#############################
+
+if [ "$BUILD_TARGET" = "darwin" ]; then
+    TARGET_DIR=Cura-${BUILD_NAME}-MacOS
+
+	rm -rf scripts/darwin/build
+	rm -rf scripts/darwin/dist
+
+	python build_app.py py2app
+	rc=$?
+	if [[ $rc != 0 ]]; then
+		echo "Cannot build app."
+		exit 1
+	fi
+
+    #Add cura version file (should read the version from the bundle with pyobjc, but will figure that out later)
+    echo $BUILD_NAME > scripts/darwin/dist/Cura.app/Contents/Resources/version
+	rm -rf CuraEngine
+	git clone ${CURA_ENGINE_REPO}
+	make -C CuraEngine
+	cp CuraEngine/CuraEngine scripts/darwin/dist/Cura.app/Contents/Resources/CuraEngine
+
+	cd scripts/darwin
+
+	# Install QuickLook plugin
+	mkdir -p dist/Cura.app/Contents/Library/QuickLook
+	cp -a STLQuickLook.qlgenerator dist/Cura.app/Contents/Library/QuickLook/
+
+	# Archive app
+	cd dist
+	$TAR cfp - Cura.app | gzip --best -c > ../../../${TARGET_DIR}.tar.gz
+	cd ..
+
+	# Create sparse image for distribution
+	hdiutil detach /Volumes/Cura\ -\ Ultimaker/
+	rm -rf Cura.dmg.sparseimage
+	hdiutil convert DmgTemplateCompressed.dmg -format UDSP -o Cura.dmg
+	hdiutil resize -size 500m Cura.dmg.sparseimage
+	hdiutil attach Cura.dmg.sparseimage
+	cp -a dist/Cura.app /Volumes/Cura\ -\ Ultimaker/Cura/
+	hdiutil detach /Volumes/Cura\ -\ Ultimaker
+	hdiutil convert Cura.dmg.sparseimage -format UDZO -imagekey zlib-level=9 -ov -o ../../${TARGET_DIR}.dmg
+	exit
+fi
+
+#############################
+# Debian .deb
+#############################
+
+if [ "$BUILD_TARGET" = "debian" ]; then
+	if [ ! -d "Power" ]; then
+		git clone https://github.com/GreatFruitOmsk/Power
+	else
+		cd Power
+		git pull
+		cd ..
+	fi
+	rm -rf CuraEngine
+	git clone ${CURA_ENGINE_REPO}
+	make -C CuraEngine
+	rm -rf scripts/linux/debian/usr/share/cura
+	mkdir -p scripts/linux/debian/usr/share/cura
+	cp -a Cura scripts/linux/debian/usr/share/cura/
+	cp -a CuraEngine/CuraEngine scripts/linux/debian/usr/share/cura/
+	cp scripts/linux/cura.py scripts/linux/debian/usr/share/cura/
+	cp -a Power/power scripts/linux/debian/usr/share/cura/
+	echo $BUILD_NAME > scripts/linux/debian/usr/share/cura/Cura/version
+	sudo chown root:root scripts/linux/debian -R
+	sudo chmod 755 scripts/linux/debian/usr -R
+	sudo chmod 755 scripts/linux/debian/DEBIAN -R
+	cd scripts/linux
+	dpkg-deb --build debian ${TARGET_DIR}.deb
+	sudo chown `id -un`:`id -gn` debian -R
+	exit
+fi
+
+#############################
+# Rest
+#############################
 
 #############################
 # Download all needed files.
@@ -91,13 +187,11 @@ if [ $BUILD_TARGET = "win32" ]; then
 	downloadURL http://ffmpeg.zeranoe.com/builds/win32/static/ffmpeg-20120927-git-13f0cd6-win32-static.7z
 	downloadURL http://sourceforge.net/projects/comtypes/files/comtypes/0.6.2/comtypes-0.6.2.win32.exe
 	downloadURL http://www.uwe-sieber.de/files/ejectmedia.zip
-	
 	#Get the power module for python
-	## ----  modificado para no utilizar github
-	#rm -rf Power
-	#git clone https://github.com/GreatFruitOmsk/Power
-	#rm -rf CuraEngine
-	#git clone ${CURA_ENGINE_REPO}
+	rm -rf Power
+	git clone https://github.com/GreatFruitOmsk/Power
+	rm -rf CuraEngine
+	git clone ${CURA_ENGINE_REPO}
 fi
 
 #############################
@@ -129,13 +223,13 @@ if [ $BUILD_TARGET = "win32" ]; then
 	mv PURELIB/OpenGL ${TARGET_DIR}/python/Lib
 	mv PURELIB/comtypes ${TARGET_DIR}/python/Lib
 	mv PLATLIB/numpy ${TARGET_DIR}/python/Lib
-	cp -r Power/power ${TARGET_DIR}/python/Lib
+	mv Power/power ${TARGET_DIR}/python/Lib
 	mv VideoCapture-0.9-5/Python27/DLLs/vidcap.pyd ${TARGET_DIR}/python/DLLs
 	mv ffmpeg-20120927-git-13f0cd6-win32-static/bin/ffmpeg.exe ${TARGET_DIR}/Cura/
 	mv ffmpeg-20120927-git-13f0cd6-win32-static/licenses ${TARGET_DIR}/Cura/ffmpeg-licenses/
 	mv Win32/EjectMedia.exe ${TARGET_DIR}/Cura/
 	
-	#rm -rf Power/
+	rm -rf Power/
 	rm -rf \$_OUTDIR
 	rm -rf PURELIB
 	rm -rf PLATLIB
@@ -166,36 +260,38 @@ cp -a Cura/* ${TARGET_DIR}/Cura
 echo $BUILD_NAME > ${TARGET_DIR}/Cura/version
 
 #add script files
-cp -a scripts/${BUILD_TARGET}/*.bat $TARGET_DIR/
+if [ $BUILD_TARGET = "win32" ]; then
+    cp -a scripts/${BUILD_TARGET}/*.bat $TARGET_DIR/
+    cp CuraEngine/CuraEngine.exe $TARGET_DIR
+else
+    cp -a scripts/${BUILD_TARGET}/*.sh $TARGET_DIR/
+fi
 
-# ---- agregando printrun ----
-cp -r scripts/${BUILD_TARGET}/Printrun $TARGET_DIR/
-cp CuraEngine/CuraEngine.exe $TARGET_DIR
-
-# borra binarios compilados
-#mingw32-make -C CuraEngine clean
-
-rm -rf scripts/win32/dist
-mv `pwd`/${TARGET_DIR} scripts/win32/dist
 #package the result
-if [ "$1" = "exe" ]; then
+if (( ${ARCHIVE_FOR_DISTRIBUTION} )); then
 	if [ $BUILD_TARGET = "win32" ]; then
 		#rm ${TARGET_DIR}.zip
 		#cd ${TARGET_DIR}
 		#7z a ../${TARGET_DIR}.zip *
 		#cd ..
 
-		if [ -f '/c/Program Files (x86)/NSIS/makensis.exe' ]; then
-			#'/c/Program Files (x86)/NSIS/makensis.exe' -DVERSION=${BUILD_NAME} 'scripts/win32/installer.nsi' >> log.txt
-			'/c/Program Files (x86)/NSIS/makensis.exe' 'scripts/win32/installer.nsi' >> log.txt
-			mv scripts/win32/Cura_Sawers.exe ./
-			echo "Cura_Sawers.exe  generado"
+		if [ ! -z `which wine` ]; then
+			#if we have wine, try to run our nsis script.
+			rm -rf scripts/win32/dist
+			ln -sf `pwd`/${TARGET_DIR} scripts/win32/dist
+			wine ~/.wine/drive_c/Program\ Files/NSIS/makensis.exe /DVERSION=${BUILD_NAME} scripts/win32/installer.nsi
+			mv scripts/win32/Cura_${BUILD_NAME}.exe ./
 		fi
+		if [ -f '/c/Program Files (x86)/NSIS/makensis.exe' ]; then
+			rm -rf scripts/win32/dist
+			mv `pwd`/${TARGET_DIR} scripts/win32/dist
+			'/c/Program Files (x86)/NSIS/makensis.exe' -DVERSION=${BUILD_NAME} 'scripts/win32/installer.nsi' >> log.txt
+			mv scripts/win32/Cura_${BUILD_NAME}.exe ./
+		fi
+	else
+		echo "Archiving to ${TARGET_DIR}.tar.gz"
+		$TAR cfp - ${TARGET_DIR} | gzip --best -c > ${TARGET_DIR}.tar.gz
 	fi
+else
+	echo "Installed into ${TARGET_DIR}"
 fi
-
-
-#rm -rf Power
-#git clone https://github.com/GreatFruitOmsk/Power
-#rm -rf CuraEngine
-#git clone https://github.com/Ultimaker/CuraEngine
